#8207 ESENER CSS: only show scroll bars when necessary
Revert the previous copy-and-paste from the customized version of
surveyanswers.css and set up the questions overlays so that the scroll
bars only appear when the overlay goes over the height of the page.

The height is actually limited by the height of
#portal-column-content, but I think that's good enough since it's the
height of the content area.

This has been tested in Firefox, Chrome, IE8,9,10